# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Homestay App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{propertyId}/rooms:
    get:
      operationId: getRooms
      tags: [rooms]
      summary: Get all rooms for a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"

    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room (host only)
      security:
        - bearerAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room (host only)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomRequest"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete room (host only)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rooms/{roomId}/availability:
    get:
      operationId: getRoomAvailability
      tags: [rooms]
      summary: Get booked dates for a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
            description: YYYY-MM format
      responses:
        "200":
          description: Booked dates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"

  # BOOKINGS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # AUTH SCHEMAS
...
    Room:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        name:
          type: string
        type:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, occupied, maintenance]
        createdAt:
          type: string
      required:
        - id
        - propertyId
        - name
        - type
        - pricePerNight
        - capacity
        - status
        - createdAt
    CreateRoomRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        photos:
          type: array
          items:
            type: string
      required:
        - name
        - type
        - pricePerNight
        - capacity
    UpdateRoomRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, occupied, maintenance]
    AvailabilityResponse:
      type: object
      properties:
        roomId:
          type: string
        bookedDates:
          type: array
          items:
            type: string
      required:
        - roomId
        - bookedDates

    # BOOKING SCHEMAS
...